requirements.txt
mysql-connector-python==9.6.0
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
pillow==12.1.1
